Reactivate Rainbow-Parenthesis.
After having a look I think that Rainbow-Parenthesis-Improved are overkill for
most programmers and you won't need it unless you do a lot of XML or HTML
editing and want the tags in rainbow.

Don't get me wrong, Rainbow-Parenthesis-Improved is an impressive feast of
software design and a powerful plug-in. I just don't that much functionality.